:ambulance: change: コンポーネント分け
